+++ CHANGELOG.md
@@ -36,6 +36,14 @@ Minimal required AGP and KGP will always be specified in README.
 
 ### Changed
 
+**BREAKING CHANGES:**
+- **android:** Minimal required AGP version is `7.4.0`
+- **android:** Default `targetSdk` changed from `32` to `33`
+- **android:** Don't set `targetSdk` in library modules. This field is deprecated and doesn't take any effect since AGP 7.4 ([b/230625468](https://issuetracker.google.com/issues/230625468#comment5))
+
+Other changes:
+- **android:** Removed workaround for [b/215407138](https://issuetracker.google.com/issues/215407138) that is fixed in AGP 7.4
+- Change target JDK for all plugins from 8 to 11
 - Update Gradle to 7.6.1
 
 ## [0.17] (2022-07-29)